PYIC-1718: Rename IPV session end lambda
