+++ lambdas/build-client-oauth-response/src/main/resources/IpvLambdaJsonLayout.json
