+++ lambdas/build-client-oauth-response/src/main/resources/log4j2.yaml
