+++ settings.gradle
@@ -2,7 +2,7 @@ rootProject.name = 'di-ipv-core-back'
 include "lib",
 		"lambdas",
 		"lambdas:issue-client-access-token",
-		"lambdas:sessionend",
+		"lambdas:build-client-oauth-response",
 		"lambdas:retrieve-cri-oauth-access-token",
 		"lambdas:get-credential-issuer-config",
 		"lambdas:evaluate-gpg45-scores",
